Update python module examples
