+++ gldcore/link/python/examples/example_4/test_memory.glm
